Add elastix support to CI (#95)
Add default elastix 4.9.0 installation during CI testing. this will let us validate the registration module with higher efficacy.

Other changes are based on the fact that we added elastix support to CI

* adding scan io tests

* try downloading elastix

* add mkdir for elastix download

* remove extra quotes

* autoformat tests

* check if elastix persists across jobs

* add elastix support - modify tests

* remove elastix version call

* update registration code

* add check if directory exists

* update testing for registration and numpy routines

* fix util temp directory, which previously caused issues with multiple test processes running in parallel

* manage paths

* autoformat registration

+++ dosma/core/registration.py
@@ -4,6 +4,7 @@ import os
 import platform
 import shutil
 import subprocess
+import sys
 import uuid
 import warnings
 from functools import partial
@@ -92,7 +93,9 @@ def register(
     assert issubclass(rtype, (Dict, Sequence))  # `rtype` must be dict or tuple
     has_output_path = bool(output_path)
     if not output_path:
-        output_path = os.path.join(env.temp_dir(), "register")
+        output_path = os.path.join(
+            env.temp_dir(), f"register-{str(uuid.uuid1())}-{str(uuid.uuid4())}"
+        )
 
     moving = [moving] if isinstance(moving, (MedicalVolume, str)) else moving
     moving_masks = (
@@ -241,7 +244,15 @@ def apply_warp(
     if rtype == str and not has_output_path:
         raise ValueError("`output_path` must be specified when `rtype=str`")
     if not output_path:
-        output_path = os.path.join(env.temp_dir(), f"apply_warp-{str(uuid.uuid1())}")
+        # TODO: Add path generation that prevents collisions during multiprocessing.
+        # When multiprocessing executes rapidly and poor seed is set, the uuids have
+        # collided. To avoid this, we append the process name to the directory that is
+        # created.
+        output_path = os.path.join(
+            env.temp_dir(), f"apply_warp-{str(uuid.uuid1())}-{str(uuid.uuid4())}"
+        )
+        if not _is_main_process():
+            output_path += mp.current_process().name
     output_path = os.path.abspath(output_path)
     os.makedirs(output_path, exist_ok=True)
 
@@ -276,7 +287,7 @@ def apply_warp(
 
     if os.path.isfile(mv_filepath):
         os.remove(mv_filepath)
-    if not has_output_path:
+    if not has_output_path and os.path.isdir(output_path):
         shutil.rmtree(output_path)
 
     return out
@@ -457,3 +468,10 @@ def _local_lib_dir():
     files = [x for x in os.listdir(fc._DOSMA_ELASTIX_FOLDER) if x.startswith("libANNlib")]
     if len(files) > 0:
         return fc._DOSMA_ELASTIX_FOLDER
+
+
+def _is_main_process():
+    py_version = tuple(sys.version_info[0:2])
+    return (py_version < (3, 8) and mp.current_process().name == "MainProcess") or (
+        py_version >= (3, 8) and mp.parent_process() is None
+    )
